+++ lib/CHANGELOG.md
@@ -4,6 +4,8 @@
  - rework README.
  - add chip path accessors to Request and Config.
  - fix clippy 1.75 lints
+ - add AsFd support to Chip and Request (bumps MSRV to 1.63).
+ - update to gpiocdev-uapi 0.6 (io safety)
 
 <a name="v0.6.0"></a>
 ## v0.6.0 - 2023-07-18